util refactoring
