+++ frontend/src/Chat/components/utils/Icon/MutedIcon.tsx
