+++ frontend/src/Chat/components/utils/Icon/privateIcon.tsx
