+++ frontend/src/Chat/components/utils/Icon/protectedIcon.tsx
